Break out dialog JSX into separate components
