+++ src/react-components/webvr-recommend-dialog.js
@@ -1,0 +1,20 @@
+import React, { Component } from "react";
+
+export default class WebVRRecommendDialog extends Component {
+  render() {
+    return (
+      <div>
+        <p>To enter Hubs with Oculus or SteamVR, you can use Firefox.</p>
+        <a className="info-dialog--action-button" href="https://www.mozilla.org/firefox">
+          Download Firefox
+        </a>
+        <p style={{ fontSize: "0.8em" }}>
+          For a full list of browsers with experimental VR support, visit{" "}
+          <a href="https://webvr.rocks" target="_blank" rel="noopener noreferrer">
+            WebVR Rocks
+          </a>.
+        </p>
+      </div>
+    );
+  }
+}
